refactor(cli,server): split the two god-functions with no hot path attached
Both are the (a) entries #1905's function-size gate listed, both move-verbatim.

imp-cli main(): 737 -> 233 code LOC. The bench, interactive and one-shot modes
move into tools/imp-cli/mode_{bench,interactive,oneshot}.cpp. This is not a new
layout - modes.h has described exactly this since mode_perplexity.cpp ("this
file is where they move as they are touched"), and main() is now load, resolve
and one dispatch. Only edits beyond the move: the 4-space dedent, a local
ImpError err per mode, params by value where the mode overrides it (interactive
raises max_tokens to 16384), and run_bench dropping the params it never read.
The three one-shot error paths that used to return straight out of main now
return an exit code, so ctx/model are freed on the way out.

handle_completions(): 593 -> 240 code LOC. The SSE loop and the blocking
response move behind a CompletionCtx, which is the shape the chat route already
has (ChatRequestContext, filled by parse_chat_request_params) - /v1/completions
was the endpoint that kept parse, stream and non-stream in one body. The moved
bodies are byte-identical apart from the dedent: each function re-aliases the 14
locals it uses instead of re-spelling them as c.x, which would have rewritten
lines nobody is changing (and `"created"` is a JSON key as well as a local).

Both [allow] entries removed - the gate now measures 233 and 240 against a hard
line of 500.

docs/BENCHMARKS.md's citation of the bench prompt ramp follows the code to
mode_bench.cpp:19; the doc-citation gate caught it, which is what it is for.
