+++ docs/audit/docs-rewrite/STALE.md
@@ -2,8 +2,8 @@
 
 Generated by scripts/docs_lint.py. Threshold 180 days.
 
-- AGENTS.md: edited 4x since the commit it says it was verified against (81ffa573)
-- CLAUDE.md: edited 7x since the commit it says it was verified against (81ffa573)
+- AGENTS.md: edited 5x since the commit it says it was verified against (81ffa573)
+- CLAUDE.md: edited 8x since the commit it says it was verified against (81ffa573)
 - CLAUDE.md:49: throughput figure with no [PROV: ...] block within 12 lines
 - CONTRIBUTING.md: edited 2x since the commit it says it was verified against (81ffa573)
 - README.md: edited 1x since the commit it says it was verified against (151ff123)
