+++ tools/imp-cli/main.cpp
@@ -349,668 +349,11 @@ int main(int argc, char** argv) {
         return rc;
     }
 
-    if (args.bench) {
-        // Synthetic benchmark mode (matches llama-bench methodology)
-        int vocab_size = imp_model_vocab_size(model);
-        std::vector<int32_t> tokens(args.bench_pp);
-        for (int i = 0; i < args.bench_pp; i++)
-            tokens[i] = i % vocab_size;
-
-        int tg_tokens = args.max_tokens;
-
-        // Greedy decode params for deterministic benchmarking
-        ImpGenerateParams bench_params = imp_generate_params_default();
-        bench_params.temperature = 0.0f;
-        bench_params.ignore_eos = 1;  // Don't stop on EOS during benchmark
-        // +1 because imp_prefill already produces the first output token;
-        // without this the request hits max_tokens one decode step early.
-        bench_params.max_tokens = tg_tokens + 1;
-
-        fprintf(stderr, "Benchmark: pp=%d, tg=%d, reps=%d\n", args.bench_pp, tg_tokens, args.bench_reps);
-
-        // Warmup: 1 full prefill+decode cycle (discarded)
-        fprintf(stderr, "Warmup...\n");
-        imp_context_reset(ctx);
-        imp_prefill_with_params(ctx, tokens.data(), args.bench_pp, &bench_params);
-        for (int s = 0; s < tg_tokens; s++) {
-            int32_t tok = 0;
-            imp_decode_step(ctx, &bench_params, &tok);
-        }
-
-        // PP benchmark
-        double pp_total_ms = 0;
-        for (int rep = 0; rep < args.bench_reps; rep++) {
-            imp_context_reset(ctx);
-            auto t0 = std::chrono::high_resolution_clock::now();
-            err = imp_prefill_with_params(ctx, tokens.data(), args.bench_pp, &bench_params);
-            auto t1 = std::chrono::high_resolution_clock::now();
-            if (err != IMP_SUCCESS) {
-                fprintf(stderr, "Prefill error on rep %d: %s\n", rep, imp_error_string(err));
-                break;
-            }
-            pp_total_ms += std::chrono::duration<double, std::milli>(t1 - t0).count();
-        }
-
-        // TG benchmark
-        double tg_total_ms = 0;
-        for (int rep = 0; rep < args.bench_reps; rep++) {
-            imp_context_reset(ctx);
-            err = imp_prefill_with_params(ctx, tokens.data(), args.bench_pp, &bench_params);
-            if (err != IMP_SUCCESS) {
-                fprintf(stderr, "Prefill error on tg rep %d: %s\n", rep, imp_error_string(err));
-                break;
-            }
-            auto t0 = std::chrono::high_resolution_clock::now();
-            for (int s = 0; s < tg_tokens; s++) {
-                int32_t tok = 0;
-                err = imp_decode_step(ctx, &bench_params, &tok);
-                if (err != IMP_SUCCESS)
-                    break;
-            }
-            auto t1 = std::chrono::high_resolution_clock::now();
-            if (err != IMP_SUCCESS) {
-                fprintf(stderr, "Decode error on rep %d: %s\n", rep, imp_error_string(err));
-                break;
-            }
-            tg_total_ms += std::chrono::duration<double, std::milli>(t1 - t0).count();
-        }
-
-        double pp_avg_ms = pp_total_ms / args.bench_reps;
-        double tg_avg_ms = tg_total_ms / args.bench_reps;
-        double pp_toks = (pp_avg_ms > 0) ? (args.bench_pp / (pp_avg_ms / 1000.0)) : 0;
-        double tg_toks = (tg_avg_ms > 0) ? (tg_tokens / (tg_avg_ms / 1000.0)) : 0;
-
-        fprintf(stderr, "pp %5d tokens  avg %8.2f ms  (%7.2f tok/s)  [%d reps]\n", args.bench_pp, pp_avg_ms,
-                pp_toks, args.bench_reps);
-        fprintf(stderr, "tg %5d tokens  avg %8.2f ms  (%7.2f tok/s)  [%d reps]\n", tg_tokens, tg_avg_ms,
-                tg_toks, args.bench_reps);
-
-        if (args.json_out)
-            imp_cli::emit_bench({resolved_model, pp_toks, tg_toks, args.bench_pp, pp_avg_ms, tg_tokens,
-                                 tg_avg_ms, args.bench_reps,
-                                 static_cast<long long>(imp::vram_own_peak_bytes() >> 20)});
-    } else if (args.interactive) {
-        // Interactive/agentic defaults to 16384 max tokens (needs headroom for
-        // long reasoning chains, code generation, and multi-step tool use)
-        if (!args.max_tokens_set) {
-            params.max_tokens = 16384;
-        }
-        // Multi-turn interactive mode using token-level API with chat template
-        imp::Tokenizer* tok = model->model->tokenizer();
-        const imp::ChatTemplate& engine_tpl = ctx->engine->chat_template();
-
-        // Resolve effective chat template: CLI override or engine-detected
-        imp::ChatTemplate chat_tpl;
-        bool have_template = false;
-
-        if (args.chat_template == "none") {
-            // No template
-        } else if (args.chat_template != "auto") {
-            // Explicit override from CLI
-            auto family = imp::ChatTemplate::parse_family(args.chat_template);
-            if (family != imp::ChatTemplateFamily::RAW) {
-                have_template = chat_tpl.init(family, *tok);
-            }
-        } else {
-            // Use engine-detected template
-            if (!engine_tpl.is_raw()) {
-                chat_tpl = engine_tpl;
-                have_template = true;
-            }
-        }
-
-        if (have_template) {
-            printf("Chat template: %s\n", imp::chat_template_family_name(chat_tpl.family()));
-        } else {
-            printf("No chat template (raw mode)\n");
-        }
-
-        printf("Interactive mode. Type 'quit' to exit.\n");
-        if (ctx->engine->has_vision()) {
-            printf("Vision enabled. Use '/image <path>' to load an image.\n");
-        }
-
-        std::vector<imp::ChatMessage> history;
-        char line[4096];
-
-        while (true) {
-            printf("\n> ");
-            fflush(stdout);
-            if (!fgets(line, sizeof(line), stdin))
-                break;
-
-            // Trim trailing newline
-            size_t len = std::strlen(line);
-            if (len > 0 && line[len - 1] == '\n')
-                line[len - 1] = '\0';
-
-            std::string input(line);
-            if (input.empty() || input == "quit" || input == "exit")
-                break;
-
-            // Handle /image command
-            if (input.rfind("/image ", 0) == 0) {
-                std::string img_path = input.substr(7);
-                // Repeating /image before sending stacks the pictures up. The
-                // pending list empties when a message consumes it, so the first
-                // /image of the next turn starts fresh without a clear command.
-                const bool first = imp_pending_image_tokens(ctx) == 0;
-                err = first ? imp_set_image(ctx, img_path.c_str()) : imp_add_image(ctx, img_path.c_str());
-                if (err != IMP_SUCCESS) {
-                    fprintf(stderr, "Error loading image: %s\n", imp_error_string(err));
-                } else {
-                    printf("Image loaded: %s (applies to your next message)\n", img_path.c_str());
-                }
-                continue;
-            }
-
-            if (have_template) {
-                // Multi-turn: append user message and apply full template
-                history.push_back({"user", input});
-                std::vector<int32_t> tokens;
-                // Same three-way split as the single-prompt path below: a
-                // pending Qwen-VL image has a count only the encoder knows, an
-                // mmproj image has a fixed one, and most turns have neither.
-                // Testing only `has_vision_input()` here — which is the mmproj
-                // tower alone — rendered a prompt with no image tokens at all,
-                // so a picture loaded with /image was never referenced and the
-                // model answered as if it had not been given one.
-                const int pending_img_tokens = imp_pending_image_tokens(ctx);
-                if (pending_img_tokens > 0) {
-                    const std::vector<int> counts = ctx->engine->pending_image_token_counts();
-                    std::string blocks;
-                    for (size_t i = 0; i < counts.size(); ++i)
-                        blocks += "<|vision_start|><|image_pad|><|vision_end|>";
-                    std::vector<imp::ChatMessage> msgs = history;
-                    msgs.back().content = blocks + msgs.back().content;
-                    tokens = chat_tpl.apply(*tok, msgs);
-                    const int32_t pad_id = tok->find_token("<|image_pad|>");
-                    const auto expanded = pad_id < 0
-                                              ? std::unexpected(std::string("tokenizer has no <|image_pad|>"))
-                                              : imp::expand_image_placeholders(tokens, pad_id, counts);
-                    if (!expanded) {
-                        fprintf(stderr, "Error placing image tokens: %s\n", expanded.error().c_str());
-                        history.pop_back();
-                        continue;
-                    }
-                } else if (ctx->engine->has_vision_input()) {
-                    tokens = chat_tpl.apply_with_image(*tok, history, 256);
-                } else {
-                    tokens = chat_tpl.apply(*tok, history);
-                }
-
-                // Reset context for fresh KV cache
-                imp_context_reset(ctx);
-
-                // Prefill with templated tokens (params apply to first sample)
-                err = imp_prefill_with_params(ctx, tokens.data(), static_cast<int>(tokens.size()), &params);
-                if (err != IMP_SUCCESS) {
-                    fprintf(stderr, "Prefill error: %s\n", imp_error_string(err));
-                    history.pop_back();
-                    continue;
-                }
-
-                // Capture the first token produced during prefill
-                // (engine->step() generates it as part of the prefill pass)
-                std::vector<int32_t> output_ids;
-                std::string response;
-                std::string interactive_text;
-                // Think-block styling: buffer output to suppress <think></think>
-                // tags and render thinking content in dim grey.
-                std::string print_buf;  // pending text not yet flushed
-
-                // Capture the first token produced during prefill
-                // (engine->step() generates it as part of the prefill pass)
-                if (ctx->active_request && !ctx->active_request->output_tokens.empty()) {
-                    int32_t first_tok = ctx->active_request->output_tokens.back();
-                    output_ids.push_back(first_tok);
-                    std::string piece = tok->decode_token(first_tok);
-                    interactive_text += piece;
-                    print_buf += piece;
-                }
-
-                // Decode token by token
-                bool in_think = false;
-                static const char* kThinkOn = "\033[2;90m";  // dim + bright black
-                static const char* kThinkOff = "\033[0m";
-
-                // Flush confirmed text from print_buf up to a safe point
-                auto flush_buf = [&]() {
-                    if (print_buf.empty())
-                        return;
-                    // Don't flush text that could be a partial tag
-                    // Max partial: "</think>" (8 chars) or "<think>" (7 chars)
-                    const size_t hold = 8;
-                    if (print_buf.size() <= hold)
-                        return;
-                    size_t safe = print_buf.size() - hold;
-                    printf("%.*s", (int)safe, print_buf.c_str());
-                    fflush(stdout);
-                    print_buf.erase(0, safe);
-                };
-
-                for (int step = 0; step < params.max_tokens; step++) {
-                    int32_t token = 0;
-                    err = imp_decode_step(ctx, &params, &token);
-                    if (err != IMP_SUCCESS)
-                        break;
-
-                    // Check stop tokens
-                    if (token == tok->eos_id())
-                        break;
-                    bool is_stop = false;
-                    for (int32_t stop_id : chat_tpl.stop_token_ids()) {
-                        if (token == stop_id) {
-                            is_stop = true;
-                            break;
-                        }
-                    }
-                    if (is_stop)
-                        break;
-
-                    output_ids.push_back(token);
-                    std::string piece = tok->decode_token(token);
-                    interactive_text += piece;
-                    print_buf += piece;
-
-                    // Scan for tag transitions in the buffer
-                    while (true) {
-                        if (!in_think) {
-                            auto pos = print_buf.find("<think>");
-                            if (pos != std::string::npos) {
-                                // Flush text before the tag normally
-                                if (pos > 0) {
-                                    printf("%.*s", (int)pos, print_buf.c_str());
-                                }
-                                // Switch to think style, consume the tag
-                                printf("%s", kThinkOn);
-                                fflush(stdout);
-                                print_buf.erase(0, pos + 7);
-                                in_think = true;
-                                continue;
-                            }
-                        } else {
-                            auto pos = print_buf.find("</think>");
-                            if (pos != std::string::npos) {
-                                // Flush thinking text before closing tag
-                                if (pos > 0) {
-                                    printf("%.*s", (int)pos, print_buf.c_str());
-                                }
-                                // Reset style, consume the tag
-                                printf("%s", kThinkOff);
-                                fflush(stdout);
-                                print_buf.erase(0, pos + 8);
-                                in_think = false;
-                                continue;
-                            }
-                        }
-                        break;
-                    }
-
-                    // Flush safe portion of buffer (keeping potential partial tags)
-                    flush_buf();
-
-                    // Check text-level stop sequences
-                    if (!args.stop_sequences.empty()) {
-                        bool text_stop = false;
-                        for (const auto& stop : args.stop_sequences) {
-                            if (interactive_text.find(stop) != std::string::npos) {
-                                text_stop = true;
-                                break;
-                            }
-                        }
-                        if (text_stop)
-                            break;
-                    }
-                }
-                // Flush remaining buffer
-                if (!print_buf.empty()) {
-                    printf("%s", print_buf.c_str());
-                }
-                if (in_think)
-                    printf("%s", kThinkOff);
-                printf("\n");
-
-                response = tok->decode(output_ids);
-                history.push_back({"assistant", response});
-            } else {
-                // Raw mode: no history, just generate
-                imp_context_reset(ctx);
-                char output[8192];
-                size_t output_len = 0;
-                err = imp_generate(ctx, input.c_str(), &params, output, sizeof(output), &output_len);
-                if (err != IMP_SUCCESS) {
-                    fprintf(stderr, "Generation error: %s\n", imp_error_string(err));
-                    continue;
-                }
-                printf("%.*s\n", (int)output_len, output);
-            }
-        }
-    } else {
-        // Single-shot mode with timing
-        if (args.prompt.empty()) {
-            fprintf(stderr, "No prompt provided. Use --prompt, --prompt-file or --interactive\n");
-        } else {
-            // Load images if specified. The first replaces whatever was
-            // pending, the rest append, so repeating --image builds a prompt
-            // with one placeholder per picture.
-            for (size_t i = 0; i < args.image_paths.size(); ++i) {
-                const char* p = args.image_paths[i].c_str();
-                err = (i == 0) ? imp_set_image(ctx, p) : imp_add_image(ctx, p);
-                if (err != IMP_SUCCESS) {
-                    fprintf(stderr, "Error loading image '%s': %s\n", p, imp_error_string(err));
-                    imp_context_free(ctx);
-                    imp_model_free(model);
-                    return imp::tools::exit_code_for(err);
-                }
-                fprintf(stderr, "Image loaded: %s\n", p);
-            }
-
-            imp::Tokenizer* tok = model->model->tokenizer();
-            const imp::ChatTemplate& engine_tpl = ctx->engine->chat_template();
-
-            // Resolve chat template
-            imp::ChatTemplate chat_tpl;
-            bool have_template = false;
-            if (args.chat_template == "none" || !params.apply_chat_template) {
-                // No template
-            } else if (args.chat_template != "auto") {
-                auto family = imp::ChatTemplate::parse_family(args.chat_template);
-                if (family != imp::ChatTemplateFamily::RAW) {
-                    have_template = chat_tpl.init(family, *tok);
-                }
-            } else if (!engine_tpl.is_raw()) {
-                chat_tpl = engine_tpl;
-                have_template = true;
-            }
-
-            // Tokenize prompt (with image tokens if vision is active)
-            std::vector<int32_t> tokens;
-            const int pending_img_tokens = imp_pending_image_tokens(ctx);
-            if (have_template && pending_img_tokens > 0) {
-                // Dynamic resolution: the template emits one <|image_pad|> per
-                // block, because the count is not knowable until the image has
-                // been resized. Render one block per image, then expand each to
-                // what its own encoder pass produced.
-                const std::vector<int> counts = ctx->engine->pending_image_token_counts();
-                std::string blocks;
-                for (size_t i = 0; i < counts.size(); ++i)
-                    blocks += "<|vision_start|><|image_pad|><|vision_end|>";
-                std::vector<imp::ChatMessage> msgs = {{"user", blocks + args.prompt}};
-                tokens = chat_tpl.apply(*tok, msgs);
-                const int32_t pad_id = tok->find_token("<|image_pad|>");
-                const auto expanded = pad_id < 0
-                                          ? std::unexpected(std::string("tokenizer has no <|image_pad|>"))
-                                          : imp::expand_image_placeholders(tokens, pad_id, counts);
-                if (!expanded) {
-                    fprintf(stderr, "Error placing image tokens: %s\n", expanded.error().c_str());
-                    imp_context_free(ctx);
-                    imp_model_free(model);
-                    return 1;
-                }
-            } else if (have_template && ctx->engine->has_vision_input()) {
-                std::vector<imp::ChatMessage> msgs = {{"user", args.prompt}};
-                tokens = chat_tpl.apply_with_image(*tok, msgs, 256);
-            } else if (have_template) {
-                std::vector<imp::ChatMessage> msgs = {{"user", args.prompt}};
-                tokens = chat_tpl.apply(*tok, msgs);
-            } else {
-                tokens = tok->encode(args.prompt);
-                // Prepend BOS when the tokenizer requires it (e.g. Gemma)
-                bool add_bos = tok->add_bos();
-                if (ctx->engine->runtime_config().generation.force_bos)
-                    add_bos = true;
-                if (add_bos) {
-                    tokens.insert(tokens.begin(), static_cast<int32_t>(tok->bos_id()));
-                }
-            }
-            int n_prompt_tokens = static_cast<int>(tokens.size());
-            if (ctx->engine->runtime_config().diagnostics.dump_tokens) {
-                fprintf(stderr, "[DUMP_TOKENS] n=%d:", n_prompt_tokens);
-                for (int ti = 0; ti < n_prompt_tokens && ti < 20; ti++)
-                    fprintf(stderr, " %d", tokens[ti]);
-                fprintf(stderr, "\n");
-            }
-
-            // Prefill with timing
-            auto t_prefill_start = std::chrono::high_resolution_clock::now();
-            err = imp_prefill_with_params(ctx, tokens.data(), n_prompt_tokens, &params);
-            auto t_prefill_end = std::chrono::high_resolution_clock::now();
-            if (err != IMP_SUCCESS) {
-                fprintf(stderr, "Prefill error: %s\n", imp_error_string(err));
-                imp_context_free(ctx);
-                imp_model_free(model);
-                return imp::tools::exit_code_for(err);
-            }
-
-            // Compute max stop length for buffering
-            size_t max_stop_len = 0;
-            for (const auto& s : args.stop_sequences)
-                max_stop_len = std::max(max_stop_len, s.size());
-
-            // Resolve think token IDs for output filtering.
-            // find_token("<think>") fails on Qwen3 BPE where <think> is a
-            // regular vocab token (ID 123649, decodes to bytes not "<think>").
-            // Fall back to encode() which handles both special and BPE tokens.
-            int32_t think_start = tok->find_token("<think>");
-            int32_t think_end = tok->find_token("</think>");
-            if (think_start < 0) {
-                auto ids = tok->encode("<think>");
-                if (ids.size() == 1) think_start = ids[0];
-            }
-            if (think_end < 0) {
-                auto ids = tok->encode("</think>");
-                if (ids.size() == 1) think_end = ids[0];
-            }
-            bool in_think = false;
-
-            auto t_decode_start = std::chrono::high_resolution_clock::now();
-            std::vector<int32_t> output_ids;
-            std::string output_text;
-            // Exactly what stdout shows, for --json (#1583). Not
-            // tok->decode(output_ids): that one carries the hidden stop and
-            // think markers, so the document would disagree with the terminal.
-            std::string visible_text;
-            if (ctx->active_request && !ctx->active_request->output_tokens.empty()) {
-                int32_t first_tok = ctx->active_request->output_tokens.back();
-                // Check stop conditions on first token
-                bool first_is_stop = (first_tok == tok->eos_id());
-                if (!first_is_stop && have_template) {
-                    for (int32_t stop_id : chat_tpl.stop_token_ids()) {
-                        if (first_tok == stop_id) {
-                            first_is_stop = true;
-                            break;
-                        }
-                    }
-                }
-                if (!first_is_stop) {
-                    output_ids.push_back(first_tok);
-                    if (think_start >= 0 && first_tok == think_start) {
-                        in_think = true;
-                    } else if (think_end >= 0 && first_tok == think_end) {
-                        in_think = false;
-                    } else if (!in_think) {
-                        std::string piece = tok->decode_token(first_tok);
-                        fprintf(stderr, "[tok=%d '%s'] ", first_tok, piece.c_str());
-                        printf("%s", piece.c_str());
-                        visible_text += piece;
-                        fflush(stdout);
-                        if (!args.stop_sequences.empty())
-                            output_text += piece;
-                    }
-                }
-            }
-
-            // Decode remaining tokens
-            for (int step = 0; step < params.max_tokens; step++) {
-                int32_t token = 0;
-                err = imp_decode_step(ctx, &params, &token);
-                if (err != IMP_SUCCESS)
-                    break;
-
-                // Hide stop tokens from the user but DON'T break the loop —
-                // the engine has the authoritative stop logic (think-state
-                // suppression, max_tokens budget). When the engine actually
-                // finishes the request the next imp_decode_step returns
-                // IMP_ERROR_INTERNAL and we exit above. Bailing here on the
-                // first eos / im_end stops generation while the engine is
-                // still inside a <think> block on Qwen3.6-NVFP4 long-context
-                // (model emits <|im_end|> after empty thought; engine flips
-                // in_think to false implicitly and continues into the actual
-                // answer; CLI was previously cutting it off mid-recovery).
-                bool hide_token = (token == tok->eos_id());
-                if (have_template && !hide_token) {
-                    for (int32_t stop_id : chat_tpl.stop_token_ids()) {
-                        if (token == stop_id) {
-                            hide_token = true;
-                            break;
-                        }
-                    }
-                }
-
-                output_ids.push_back(token);
-                if (think_start >= 0 && token == think_start) {
-                    in_think = true;
-                    hide_token = true;
-                } else if (think_end >= 0 && token == think_end) {
-                    in_think = false;
-                    hide_token = true;
-                } else if (in_think) {
-                    hide_token = true;
-                }
-                std::string piece = tok->decode_token(token);
-                if (step < 10)
-                    fprintf(stderr, "[tok=%d '%s'] ", token, piece.c_str());
-                if (!hide_token) {
-                    printf("%s", piece.c_str());
-                    visible_text += piece;
-                    fflush(stdout);
-                }
-
-                // Check text-level stop sequences
-                if (!args.stop_sequences.empty()) {
-                    output_text += piece;
-                    bool stop_found = false;
-                    for (const auto& stop : args.stop_sequences) {
-                        if (output_text.find(stop) != std::string::npos) {
-                            stop_found = true;
-                            break;
-                        }
-                    }
-                    if (stop_found)
-                        break;
-                }
-            }
-            auto t_decode_end = std::chrono::high_resolution_clock::now();
-            printf("\n");
-
-            int n_output_tokens = static_cast<int>(output_ids.size());
-            double prefill_ms =
-                std::chrono::duration<double, std::milli>(t_prefill_end - t_prefill_start).count();
-            double decode_ms =
-                std::chrono::duration<double, std::milli>(t_decode_end - t_decode_start).count();
-            double total_ms =
-                std::chrono::duration<double, std::milli>(t_decode_end - t_prefill_start).count();
-
-            double pp_toks = (prefill_ms > 0) ? (n_prompt_tokens / (prefill_ms / 1000.0)) : 0;
-            double tg_toks = (decode_ms > 0 && n_output_tokens > 1)
-                                 ? ((n_output_tokens - 1) / (decode_ms / 1000.0))
-                                 : 0;
-
-            fprintf(stderr, "\n");
-            fprintf(stderr, "pp %5d tokens in %8.2f ms  (%7.2f tok/s)\n", n_prompt_tokens, prefill_ms,
-                    pp_toks);
-            fprintf(stderr, "tg %5d tokens in %8.2f ms  (%7.2f tok/s)\n", n_output_tokens, decode_ms,
-                    tg_toks);
-            fprintf(stderr, "total   %8.2f ms\n", total_ms);
-
-            // The generated text is on stderr with everything else under
-            // --json, so it has to come back in the document or the mode
-            // would silently swallow the answer.
-            if (args.json_out)
-                imp_cli::emit_generate({resolved_model, visible_text, n_prompt_tokens, n_output_tokens,
-                                        pp_toks, tg_toks, prefill_ms, decode_ms, total_ms});
-
-            // Phase 3.5 telemetry: report MTP draft accuracy if measured.
-            if (imp::Engine* engine = ctx->engine.get(); engine && engine->mtp_spec_decode_enabled()) {
-                auto acc = engine->mtp_accuracy();
-                if (acc.total > 0) {
-                    fprintf(stderr, "mtp     %d / %d drafts matched (%.1f%% accept rate)\n",
-                            acc.matches, acc.total, 100.0f * acc.rate());
-                }
-                // Per-lookahead chain accept (K>1 measurement). chain_accept_[0]
-                // duplicates mtp accuracy above; print [1..] only when present.
-                auto chain = engine->mtp_chain_accept();
-                for (size_t k = 1; k < chain.size(); ++k) {
-                    if (chain[k].total > 0) {
-                        fprintf(stderr,
-                                "mtp[k=%zu] %d / %d drafts matched (%.1f%% accept @ +%zu lookahead)\n",
-                                k, chain[k].matches, chain[k].total, 100.0f * chain[k].rate(), k);
-                    }
-                }
-
-                // Stage 0 tree-ceiling table: per-depth top-w hit rate.
-                // depth d=k+1; width w=1..W. lookahead-0 (depth 1) is
-                // teacher-forced; depth ≥2 is self-chained (lower bound).
-                auto cw = engine->mtp_chain_accept_width();
-                constexpr int W = imp::Engine::kMtpMeasureW;
-                if (!cw.empty()) {
-                    fprintf(stderr, "\n-- MTP tree-ceiling probe (top-w hit rate per depth) --\n");
-                    fprintf(stderr, "depth   n  ");
-                    for (int w = 0; w < W; ++w) fprintf(stderr, "  top-%d", w + 1);
-                    fprintf(stderr, "   (depth1=teacher-forced, depth>=2=self-chained lower bound)\n");
-                    for (size_t k = 0; k < cw.size(); ++k) {
-                        if (cw[k].total == 0) continue;
-                        fprintf(stderr, "  %2zu %5d  ", k + 1, cw[k].total);
-                        for (int w = 0; w < W; ++w)
-                            fprintf(stderr, " %5.1f%%", 100.0f * cw[k].rate(w));
-                        fprintf(stderr, "\n");
-                    }
-                    // Derived expected accept length (tokens emitted per verify):
-                    // E[accept] = sum_{d>=1} prod_{j=1..d} p(j), with the bonus
-                    // token = +1. Linear uses top-1; tree uses top-w per depth.
-                    for (int w = 0; w < W; ++w) {
-                        double e = 0.0, prod = 1.0;
-                        for (size_t k = 0; k < cw.size(); ++k) {
-                            if (cw[k].total == 0) break;
-                            prod *= cw[k].rate(w);
-                            e += prod;
-                        }
-                        fprintf(stderr, "  E[accept] top-%d = %.3f draft tokens%s\n",
-                                w + 1, e, w == 0 ? " (linear baseline)" : "");
-                    }
-                }
-            }
-
-            // Benchmark using Engine::generate() (conditional graph loop) for comparison.
-            // This eliminates per-step host overhead — shows true GPU-limited throughput.
-            if (ctx->engine->runtime_config().bench.generate) {
-                // Reset context for fresh generation
-                imp_context_reset(ctx);
-
-                // Use Engine::generate() directly for accurate timing
-                imp::Engine* engine = ctx->engine.get();
-                auto t_gen_start = std::chrono::high_resolution_clock::now();
-                std::string gen_result = engine->generate(args.prompt, params.max_tokens, params.temperature,
-                                                          params.top_p, params.top_k, params.seed,
-                                                          have_template);
-                auto t_gen_end = std::chrono::high_resolution_clock::now();
-
-                // Count output tokens by encoding the result
-                auto gen_toks = tok->encode(gen_result);
-                int gen_n = static_cast<int>(gen_toks.size());
-                double gen_total_ms =
-                    std::chrono::duration<double, std::milli>(t_gen_end - t_gen_start).count();
-                // Estimate decode time: total - prefill (reuse prefill timing from above)
-                double gen_decode_ms = gen_total_ms - prefill_ms;
-                double gen_toks_s = (gen_decode_ms > 0 && gen_n > 0) ? (gen_n / (gen_decode_ms / 1000.0)) : 0;
-                fprintf(stderr, "graph-loop: %d tg tokens in %.2f ms (%.2f tok/s, %.2f ms total)\n", gen_n,
-                        gen_decode_ms, gen_toks_s, gen_total_ms);
-            }
-        }
-    }
+    const int rc = args.bench         ? imp_cli::run_bench(ctx, model, args, resolved_model)
+                   : args.interactive ? imp_cli::run_interactive(ctx, model, args, params)
+                                      : imp_cli::run_oneshot(ctx, model, args, params, resolved_model);
 
     imp_context_free(ctx);
     imp_model_free(model);
-    return 0;
+    return rc;
 }
